TST: split up test_merge
